I finished the auth func!
