+++ Lesson7v3/db.json
@@ -1,0 +1,8 @@
+{
+  "Feityanka": "12345",
+  "Qasper": "makarena1234",
+  "Reyneria": "adeptusxiao",
+  "Loh": "123",
+  "Chmonya": "loh",
+  "Pipka": "a?"
+}
